+++ migrations/1771651931.sh
@@ -1,4 +1,8 @@
 echo "Hide wiremix and limine-snapper-restore from app launcher"
 
-cp $OMYBUNTU_PATH/applications/hidden/wiremix.desktop ~/.local/share/applications/
-cp $OMYBUNTU_PATH/applications/hidden/limine-snapper-restore.desktop ~/.local/share/applications/
+if [[ -f $OMYBUNTU_PATH/applications/hidden/wiremix.desktop ]]; then
+  cp $OMYBUNTU_PATH/applications/hidden/wiremix.desktop ~/.local/share/applications/
+fi
+if [[ -f $OMYBUNTU_PATH/applications/hidden/limine-snapper-restore.desktop ]]; then
+  cp $OMYBUNTU_PATH/applications/hidden/limine-snapper-restore.desktop ~/.local/share/applications/
+fi
